redirects for agon-docs
adds in index files for the agon-docs site structure, with redirects for all pages

this should let bookmarks and references to the old site automatically get redirected to the new one

the redirects represent the docs site structure as of 19 April 2025, with one adjustment; the `MOS-API` page is being redirected to `MOS/API`.  that page path change will happen when the latest documentation update is merged, which will happen shortly after the organisation rename.

whilst other redirects could be added to allow some other paths that _had_ existed in the recent past, as this site has a 404.html to redirect to the root of the docs site that is not strictly necessary

this commit is pre-organisation rename, so for testing to verify this works all URLs are currently pointing to agonconsole8.github.io - after the organisation rename has happened these URLs will be switched, and the repository renamed, so that URLs at the agonconsole8.github.io will redirect to agonplatform.github.io